+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-choropleth-division/map-choropleth-division.component.html
