+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-choropleth-division/map-choropleth-division.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { MapClorophletDivisionComponent } from './map-clorophlet-division.component';
+import { MapClorophletDivisionComponent } from './map-choropleth-division.component';
 
 describe('MapClorophletDivisionComponent', () => {
   let component: MapClorophletDivisionComponent;
