+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-choropleth-division/map-choropleth-division.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MapClorophletDivisionComponent } from './map-clorophlet-division.component';
+import { MapClorophletDivisionComponent } from './map-choropleth-division.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
 import { MatDialogModule } from '@angular/material/dialog';
