+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-choropleth/map-choropleth.component.html
