+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-choropleth/map-choropleth.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MapClorophletComponent } from './map-clorophlet.component';
+import { MapClorophletComponent } from './map-choropleth.component';
 import { MatDialogModule } from '@angular/material/dialog';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -11,7 +11,7 @@ import { MatTableModule } from '@angular/material/table';
 import { SafeButtonModule } from '../../../../ui/button/button.module';
 import { MatSliderModule } from '@angular/material/slider';
 import { MatSelectModule } from '@angular/material/select';
-import { MapClorophletDivisionModule } from '../map-clorophlet-division/map-clorophlet-division.module';
+import { MapClorophletDivisionModule } from '../map-choropleth-division/map-choropleth-division.module';
 
 /**
  * Single Clorophlet Configuration in Map Settings Module.
