+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-choropleths/map-choropleths.component.scss
