+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-choropleths/map-choropleths.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MapClorophletsComponent } from './map-clorophlets.component';
+import { MapClorophletsComponent } from './map-choropleths.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
 import { SafeButtonModule } from '../../../../ui/button/button.module';
 import { SafeIconModule } from '../../../../ui/icon/icon.module';
 import { SafeAlertModule } from '../../../../ui/alert/alert.module';
 import { MatTableModule } from '@angular/material/table';
-import { MapClorophletModule } from '../map-clorophlet/map-clorophlet.module';
+import { MapClorophletModule } from '../map-choropleth/map-choropleth.module';
 
 /**
  * List of clorophlets in Map Settings Module.
