+++ projects/safe/src/lib/components/widgets/map-settings/map-layers/map-layers.module.ts
@@ -5,7 +5,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatTabsModule } from '@angular/material/tabs';
 import { TranslateModule } from '@ngx-translate/core';
 import { MapMarkersModule } from './map-markers/map-markers.module';
-import { MapClorophletsModule } from './map-clorophlets/map-clorophlets.module';
+import { MapClorophletsModule } from './map-choropleths/map-choropleths.module';
 import { MapOnlineLayersModule } from './map-online-layers/map-online-layers.module';
 
 /**
